Replace homebrew with shell function
In order to keep it simple, this commit replaces homebrew with a shell
function that keeps track of every package we'd like installed. At the
end of setup, so long as we are on a linux host, we do the actual
installing with the list of packages provided as we did each tag setup.

The major reason to not use homebrew on linux was consistency. Some
packages instantly worked great while others refused to actually work
once installed. Additionally, size was a major concern as the amount of
space consumed by homebrew on linux was incredibly large in comparison
to just apt-get install <pkgname>.

Finally, this commit updates the Dockerfile to use a different base
image. This base image is still Ubuntu, just a newer version and with
all of the build/compile dependencies already installed. This is to help
speed up cold docker image builds.