Update nvchad setup
In order to keep pace with the latest changes to nvchad, this commit
updates most of the neovim configuration to adhere to how nvchad wants
it organized for the v2.0 version. It also adds in fish shell
configuration as it seems more batteries included and faster (compared
to zsh with antigen and oh-my-zsh).

Additionally, this commit updates the Dockerfile further to allow a full
build of the new setup without installing packages via the setup script.
This is in order to ensure we can cache these package installs earlier
in the Dockerfile and enjoy more caching when building new images.