Fix a few files for general linux compatibility
In order to ensure that setup/setup runs properly on an Ubuntu Linux
machine, this commit adds a check for OSX before running anything
homebrew related. Additionally, it removes docker packages as these are
not available via standard apt repositories and require specific sources
added.

Finally, this commit adjusts the crates.nvim plugin for neovim to load
when we open a Cargo.toml file and immediately show the version
information and make keybinds available.