Update setup scripts and nvim tag
In order to ensure my setup scripts work properly even if a specific
step fails to complete, this commit removes most of the `set -e` calls
in the zsh setup scripts. This means any failures will be reported
during the run but everything will continue on. We want this behavior as
it means a more consistent setup after a single run and any errors that
occur during the run can be addressed after.

Additionally, this commit updates the neovim tag to include more of my
recent settings.