Refactor databases into their own modules
In order to keep database configuration logically grouped, yet allow
each database to have it's own configuration, this commit refactors the
databases.nix module to just be an entrypoint that includes the other
database modules now. If we need to extend this to include other
databases, their config would now live in it's own module.