Add php blade template support
In order to ensure there is tree-sitter powered syntax highlighting
in php blade templates, this commit adds configuration for tree-sitter
queries to extend the default blade/php configuration with a few more
directives that make all of the languages in a php blade template work.
Additionally, this commit makes intellephense work in a blade file as
well.