Reorganize packages to be more consistent
In order to make more things correctly shared across setups, this commit
moves a lot of neovim specific packages into the neovim module and out
of other places. This keeps the list of packages more focused on the
intent of the module providing them. Additionally, this commit is
updating my fish configuration so my colors persist across machines and
aren't hidden in the gitignored fish_variables. And finally, this commit
adds setup for ruby-lsp and nil_ls to make sure these language servers
are set up how I want them.