Add new git aliases and refactor existing ones
In order to make the fixup/squash workflow easier, this commit adds two
new aliases to my fish configuration for --autosquash local and
--autosquash with remote. This let's me chose to rebase with my local
copy of the default branch or the remote copy of the default branch.

Additionally, this commit adjusts a few neovim settings for lsp stuff.
And finally, it adds mkcert for local self signed ssl certs and managing
trusting them easily.