Refactor nvim plugins and config directory
In order to simplify my neovim setup, this commit removes the configs
directory and instead puts the config for each plugin along with the
lazyspec for that plugin in the same file located in the plugins
directory. Additionally, it refactors the lsp setup to reduce code
duplication and make adding new language servers simpler.